Rename `/planbudget` to `/plan`
