+++ src/features/plan/components/Budget.js
@@ -1,4 +1,4 @@
-import { useState, useContext } from "react";
+import { useContext, useState } from "react";
 import ViewBudget from "./ViewBudget";
 import EditBudget from "./EditBudget";
 import { AppContext } from "../context/BudgetPlannerContext";
