+++ src/features/plan/components/EditBudget.js
@@ -1,11 +1,11 @@
 import React, { useState } from "react";
 import {
     Box,
-    Container,
     Button,
-    Spacer,
+    Container,
     HStack,
     Input,
+    Spacer,
     Text,
 } from "@chakra-ui/react";
