+++ src/features/plan/components/ExpenseList.js
@@ -1,4 +1,4 @@
-import React, { useContext, useState, useEffect } from "react";
+import React, { useContext, useEffect, useState } from "react";
 import ExpenseItem from "./ExpenseItem";
 import { AppContext } from "../context/BudgetPlannerContext";
 import { Input, StackDivider, VStack } from "@chakra-ui/react";
